Convert int to size_t in fpdf_sysfontinfo.cpp
Then widen all the way down into called functions.

-- Fix some pre-existing IWYU.
-- Make one format specifier unsigned

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/91130
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fxge/cfx_folderfontinfo.cpp
@@ -359,9 +359,9 @@ void* CFX_FolderFontInfo::GetFont(const ByteString& face) {
   return it != m_FontList.end() ? it->second.get() : nullptr;
 }
 
-uint32_t CFX_FolderFontInfo::GetFontData(void* hFont,
-                                         uint32_t table,
-                                         pdfium::span<uint8_t> buffer) {
+size_t CFX_FolderFontInfo::GetFontData(void* hFont,
+                                       uint32_t table,
+                                       pdfium::span<uint8_t> buffer) {
   if (!hFont)
     return 0;
